DataTableTest renamed DataTable (original DataTable deleted to avoid duplicata) + input header set opacity styled, not displayed anymore
